Changed way of informing 'not consumed' in Embedded Jetty
