+++ src/main/java/spark/webserver/JettyHandler.java
@@ -16,23 +16,17 @@
  */
 package spark.webserver;
 
-import java.io.ByteArrayInputStream;
 import java.io.IOException;
 
 import javax.servlet.Filter;
-import javax.servlet.ReadListener;
 import javax.servlet.ServletException;
-import javax.servlet.ServletInputStream;
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletRequestWrapper;
 import javax.servlet.http.HttpServletResponse;
 
 import org.eclipse.jetty.server.Request;
 import org.eclipse.jetty.server.session.SessionHandler;
 import org.eclipse.jetty.util.log.Log;
 import org.eclipse.jetty.util.log.Logger;
-
-import spark.utils.IOUtils;
 
 /**
  * Simple Jetty Handler
@@ -56,60 +50,15 @@ public class JettyHandler extends SessionHandler {
             HttpServletRequest request,
             HttpServletResponse response) throws IOException, ServletException {
 
-        try {
-            // wrap the request so 'getInputStream()' can be called multiple times
-            filter.doFilter(new HttpRequestWrapper(request), response, null);
-            baseRequest.setHandled(true);
-        } catch (NotConsumedException ignore) {
-            // TODO : Not use an exception in order to be faster.
+        HttpRequestWrapper wrapper = new HttpRequestWrapper(request);
+        filter.doFilter(wrapper, response, null);
+
+        if (wrapper.notConsumed()) {
             baseRequest.setHandled(false);
+        } else {
+            baseRequest.setHandled(true);
         }
+
     }
 
-    private static class HttpRequestWrapper extends HttpServletRequestWrapper {
-        private byte[] cachedBytes;
-
-        public HttpRequestWrapper(HttpServletRequest request) {
-            super(request);
-        }
-
-        @Override
-        public ServletInputStream getInputStream() throws IOException {
-            if (cachedBytes == null) {
-                cacheInputStream();
-            }
-            return new CachedServletInputStream();
-        }
-
-        private void cacheInputStream() throws IOException {
-            cachedBytes = IOUtils.toByteArray(super.getInputStream());
-        }
-
-        private class CachedServletInputStream extends ServletInputStream {
-            private ByteArrayInputStream byteArrayInputStream;
-
-            public CachedServletInputStream() {
-                byteArrayInputStream = new ByteArrayInputStream(cachedBytes);
-            }
-
-            @Override
-            public int read() {
-                return byteArrayInputStream.read();
-            }
-
-            @Override
-            public boolean isFinished() {
-                return byteArrayInputStream.available() <= 0;
-            }
-
-            @Override
-            public boolean isReady() {
-                return byteArrayInputStream.available() >= 0;
-            }
-
-            @Override
-            public void setReadListener(ReadListener readListener) {
-            }
-        }
-    }
 }
