--- src/main/java/spark/webserver/NotConsumedException.java
@@ -1,28 +1,0 @@
-/*
- * Copyright 2011- Per Wendel
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *  
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package spark.webserver;
-
-/**
- * Exception used for stopping the execution
- *
- * @author Yan Bonnel
- */
-public class NotConsumedException extends RuntimeException {
-    public static final NotConsumedException DEFAULT = new NotConsumedException();
-
-    private static final long serialVersionUID = 6748899505989134484L;
-}
